Remoção de coisas desenecessárias.
