--- Ateliex.DTO/Ateliex.DTO.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.1</TargetFramework>
-    <RootNamespace>Ateliex</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Folder Include="Cadastro\Modelos\" />
-    <Folder Include="Decisoes\Comerciais\" />
-  </ItemGroup>
-
-</Project>
