+++ Ateliex.DomainDrivenDesign.sln
@@ -13,7 +13,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "3 - Camada de Domínio", "3
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "4 - Camada de Infraestrutura", "4 - Camada de Infraestrutura", "{5AB1DED7-0B42-4C4F-8225-C00571B91560}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "5 - Documentação", "5 - Documentação", "{334A6437-C605-4DCE-9262-D19C0C1A36B2}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "6 - Documentação", "6 - Documentação", "{334A6437-C605-4DCE-9262-D19C0C1A36B2}"
 	ProjectSection(SolutionItems) = preProject
 		README.md = README.md
 	EndProjectSection
@@ -36,17 +36,15 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ateliex.Domain", "Ateliex.D
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ateliex.Application", "Ateliex.Application\Ateliex.Application.csproj", "{D76CB69D-54B1-4F22-A66C-E4CC91E320DA}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ateliex.EntityFrameworkCore", "Ateliex.EntityFrameworkCore\Ateliex.EntityFrameworkCore.csproj", "{3C3FB54C-E844-45AB-960D-8373A14646FD}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ateliex.Net.Http", "Ateliex.Net.Http\Ateliex.Net.Http.csproj", "{ECD8D626-23FB-44D1-9246-F6A71B5E0207}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ateliex.Web", "Ateliex.Web\Ateliex.Web.csproj", "{4C5CC470-685D-4153-803A-03364F544138}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ateliex.DTO", "Ateliex.DTO\Ateliex.DTO.csproj", "{3B5E830D-66E4-4815-A59F-4831755C2B9B}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ateliex.Data.Sqlite", "Ateliex.Data.Sqlite\Ateliex.Data.Sqlite.csproj", "{573BD99D-7CE4-42F3-880E-940F1269CA13}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ateliex.Data.SqlClient", "Ateliex.Data.SqlClient\Ateliex.Data.SqlClient.csproj", "{4A5B99F9-BFE3-42E9-8437-A0F613BD29A8}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "5 - Dependências Externas", "5 - Dependências Externas", "{CB9142A7-683A-46D6-A62F-FCC1947ABDBD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -82,10 +80,6 @@ Global
 		{D76CB69D-54B1-4F22-A66C-E4CC91E320DA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D76CB69D-54B1-4F22-A66C-E4CC91E320DA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D76CB69D-54B1-4F22-A66C-E4CC91E320DA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3C3FB54C-E844-45AB-960D-8373A14646FD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3C3FB54C-E844-45AB-960D-8373A14646FD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3C3FB54C-E844-45AB-960D-8373A14646FD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3C3FB54C-E844-45AB-960D-8373A14646FD}.Release|Any CPU.Build.0 = Release|Any CPU
 		{ECD8D626-23FB-44D1-9246-F6A71B5E0207}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{ECD8D626-23FB-44D1-9246-F6A71B5E0207}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{ECD8D626-23FB-44D1-9246-F6A71B5E0207}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -94,10 +88,6 @@ Global
 		{4C5CC470-685D-4153-803A-03364F544138}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4C5CC470-685D-4153-803A-03364F544138}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4C5CC470-685D-4153-803A-03364F544138}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3B5E830D-66E4-4815-A59F-4831755C2B9B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3B5E830D-66E4-4815-A59F-4831755C2B9B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3B5E830D-66E4-4815-A59F-4831755C2B9B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3B5E830D-66E4-4815-A59F-4831755C2B9B}.Release|Any CPU.Build.0 = Release|Any CPU
 		{573BD99D-7CE4-42F3-880E-940F1269CA13}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{573BD99D-7CE4-42F3-880E-940F1269CA13}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{573BD99D-7CE4-42F3-880E-940F1269CA13}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -121,10 +111,8 @@ Global
 		{13CD1CA2-5A27-4804-8AC8-DCB9584AF3C7} = {A9924D55-581E-49F7-AF6A-E6E797FC6A7A}
 		{B636A62D-DC67-4525-B192-CF3941F40576} = {C045CE37-7E78-429E-9265-2956D1787A1D}
 		{D76CB69D-54B1-4F22-A66C-E4CC91E320DA} = {A41C1300-B8AB-49C0-BB41-ECC7673B32A5}
-		{3C3FB54C-E844-45AB-960D-8373A14646FD} = {89E4BB85-AE46-4E96-93C7-269794163702}
 		{ECD8D626-23FB-44D1-9246-F6A71B5E0207} = {89E4BB85-AE46-4E96-93C7-269794163702}
 		{4C5CC470-685D-4153-803A-03364F544138} = {BAEF6DF0-4906-4E64-9158-B00BDC3A0BD2}
-		{3B5E830D-66E4-4815-A59F-4831755C2B9B} = {89E4BB85-AE46-4E96-93C7-269794163702}
 		{573BD99D-7CE4-42F3-880E-940F1269CA13} = {89E4BB85-AE46-4E96-93C7-269794163702}
 		{4A5B99F9-BFE3-42E9-8437-A0F613BD29A8} = {89E4BB85-AE46-4E96-93C7-269794163702}
 	EndGlobalSection
