+++ Ateliex.Net.Http/Ateliex.Net.Http.csproj
@@ -11,8 +11,4 @@
     <ProjectReference Include="..\System.Extensions\System.Extensions.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Decisoes\Comerciais\" />
-  </ItemGroup>
-
 </Project>
