+++ Ateliex.Windows.Infrastructure/DbModule.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+using System.DomainModel;
+
+namespace Ateliex
+{
+    public static class DbModule
+    {
+        internal static IServiceCollection AddDbServices(this IServiceCollection services)
+        {
+            services.AddTransient<IAppendOnlyStore>(factory => new SqliteStore(@"Data Source=Ateliex.db"));
+
+            return services;
+        }
+    }
+}
