+++ Ateliex.Windows.Infrastructure/HttpModule.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace Ateliex
+{
+    public static class HttpModule
+    {
+        internal static IServiceCollection AddHttpServices(this IServiceCollection services)
+        {
+            return services;
+        }
+    }
+}
