+++ System.Extensions/ComponentModel/EnumDescriptionTypeConverter.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Reflection;
-using System.Text;
+﻿using System.Reflection;
 
 namespace System.ComponentModel
 {
