+++ System.Xaml.Extensions/EnumBindingSourceExtension.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Windows.Markup;
+﻿using System.Windows.Markup;
 
 namespace System
 {
